# contract/swagger.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: "NEO tracker"
  description: A sample API for tracking near Earth objects
  contact:
    name: Denis Zatsepin
    email: denis@zatsepin.spb.ru
  license:
    name: MIT
    url: http://github.com/gruntjs/grunt/blob/master/LICENSE-MIT
host: localhost:8080
basePath: /api
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
paths:
  /neo/hazardous:
    x-swagger-router-controller: neo.controller
    get:
      operationId: getHazardous
      tags:
        - neo
      description: Returns potentially hazardous asteroids
      responses:
        200:
          description: neo response
          schema:
            type: array
            items:
              $ref: '#/definitions/Neo'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'

  /neo/fastest:
    x-swagger-router-controller: neo.controller
    get:
      operationId: getFastest
      tags:
        - neo
      parameters:
        - name: hazardous
          in: query
          description: Filter potentially hazardous asteroids
          required: false
          type: boolean
          default: false
      description: Returns fastest asteroids
      responses:
        200:
          description: neo response
          schema:
            $ref: '#/definitions/Neo'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'

  /neo/best-year:
    x-swagger-router-controller: neo.controller
    get:
      operationId: getBestYear
      tags:
        - neo
      parameters:
        - name: hazardous
          in: query
          description: Filter potentially hazardous asteroids
          required: false
          type: boolean
          default: false
      description: Returns a year with most asteroids
      responses:
        200:
          description: neo response
          schema:
            $ref: '#/definitions/Year'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'

  /neo/best-month:
    x-swagger-router-controller: neo.controller
    get:
      operationId: getBestMonth
      tags:
        - neo
      parameters:
        - name: hazardous
          in: query
          description: Filter potentially hazardous asteroids
          required: false
          type: boolean
          default: false
      description: Returns a year with most asteroids
      responses:
        200:
          description: neo response
          schema:
            $ref: '#/definitions/Month'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/Error'

definitions:
  Neo:
    properties:
      date:
        type: string
      reference:
        type: string
      name:
        type: string
      speed:
        type: number 
      isHazardous:
        type: boolean
  Year:
    properties:
      year:
        type: number 
      count:
        type: number
  Month:
    properties:
      month:
        type: number 
      count:
        type: number
  Error:
    required:
      - error
      - message
    properties:
      error:
        type: string
      message:
        type: string